requirements.txt
opencv-python==4.8.0.76
numpy==1.26.0
Pillow==10.0.0
openai==2.28.0
gradio==4.44.0
python-dotenv==1.0.0
tqdm==4.65.0
requests==2.28.0
imageio-ffmpeg==0.4.9
faster-whisper==1.2.1
funasr==1.3.1
modelscope==1.9.0
torchaudio==2.10.0
